enable auto migration for pro(production) mode

+++ conf/db/migrations/0004_init_permission_data.down.sql
@@ -1,2 +1,2 @@
-delete from role_permission;
+delete from role_permission where role_id = 1;
 delete from permission;
